pyproject.toml
[project]
name = "download-webpage-data"
version = "0.1.0"
description = "A tool to download website contents through Tor with German exit nodes"
dependencies = [
    "requests>=2.31.0",
    "PySocks>=1.7.1",
    "stem>=1.8.2",
    "beautifulsoup4>=4.12.3",
    "urllib3>=2.2.0",
    "wget>=3.2"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/download_webpage_data"]
